+++ starkbank/merchantpurchase/__merchantpurchase.py
@@ -59,7 +59,7 @@ def get(id, user=None):
     return rest.get_id(resource=_resource, id=id, user=user)
 
 
-def query(limit=None, after=None, before=None, status=None, tags=None, ids=None, user=None):
+def query(limit=None, after=None, before=None, status=None, tags=None, ids=None, holder_id=None, user=None):
     return rest.get_stream(
         resource=_resource,
         limit=limit,
@@ -68,11 +68,12 @@ def query(limit=None, after=None, before=None, status=None, tags=None, ids=None,
         status=status,
         tags=tags,
         ids=ids,
+        holder_id=holder_id,
         user=user,
     )
 
 
-def page(cursor=None, limit=None, after=None, before=None, status=None, tags=None, ids=None, user=None):
+def page(cursor=None, limit=None, after=None, before=None, status=None, tags=None, ids=None, holder_id=None, user=None):
     return rest.get_page(
         resource=_resource,
         cursor=cursor,
@@ -82,6 +83,7 @@ def page(cursor=None, limit=None, after=None, before=None, status=None, tags=Non
         status=status,
         tags=tags,
         ids=ids,
+        holder_id=holder_id,
         user=user,
     )